requirements.txt
streamlit
streamlit-lottie
PyMuPDF
Pillow
requests
jsonschema